refactor and unit test
