+++ tests/CMakeLists.txt
@@ -17,6 +17,7 @@ endif()
 set(test_srcs ${test_srcs} test-core.cc)
 set(test_srcs ${test_srcs} test-diskio.cc)
 set(test_srcs ${test_srcs} test-fs.cc)
+set(test_srcs ${test_srcs} test-hsv.cc)
 
 add_executable(test-conky test-common.cc ${test_srcs})
 target_link_libraries(test-conky conky_core)
